+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [2.0.3] - Unreleased
+### Added
+- add critical log on rare case where developer forget to export the Site instance in the `_config.ts`
+
 ### Changed
 - `decap_cms` plugin: Add a script in the homepage to redirect to /admin/
   when an invite token or recovery token is detected from netlify identity.